neopin: split into two adapters
